Issue #38: Split oversized test file to comply with 300 LOC limit
- Replaced test_merge_panel_data.py (745 LOC) with two focused files
- test_panel_units.py: mock-data unit tests for session_mapping and load_emotion_data
- test_panel_integration.py: real-data integration tests for full merge pipeline
- All 5 modified files now pass check_python_standards.py
